[sinttest] Refactor MUC tests ready to add more
